Maybe introduce VirtualizedVolumes?
This does have an itch I'd like to scratch, which includes a renaming
of some Mutable as Modifiable, and then having the Mutable variants
being more easily factory made. The important addition here is that
the virtual volumes have a `realize` function that accepts a RegistryHolder
to "convert". This may become useful for Schematics/Archetype volumes
as they are the primary consumers of having literally only ResourceKeys
for all non-containered data, like BlockStates, Biomes, etc.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>